Bug Fix
Today, I ran the program and saw that there were two wind speeds. When I inspected the website I searched for the second wind speed and saw that it was from a hurricane. The hurricane wind speed was not showing up previously as it was not happening the other day. The data is pulled by referencing the class and it appears that this site reused the class name and added an additional index. This additional index was being pulled from my .text() function because the default option pulls every index under that class name. To avoid this happening in the future, I specified that I only want index 0 for each class that is referenced except for sunrise and sunset. Earlier, I had used StringBuilder to separate the sunrise and sunset because I did not know they were on separate indexes so now they do not need to be split.